+++ matchms/importing/load_from_json.py
@@ -112,3 +112,12 @@ def dict2spectrum(spectrum_dict: dict,
                         metadata_harmonization=metadata_harmonization)
     logger.info("Empty spectrum found (no peaks in 'peaks_json'). Will not be imported.")
     return None
+
+
+def scores_json_decoder(dct):
+    """
+    Object_hook function to convert JSON dictionary with :py:class:`~matchms.Score.Score` object into a python dictionary.
+    """
+    if "__Scores__" not in dct and "__Similarity__" not in dct:
+        return dict2spectrum(dct, metadata_harmonization=False)
+    return dct